feat: add corresponding test

+++ src/UsdnProtocol/UsdnProtocolActions.sol
@@ -13,7 +13,7 @@ abstract contract UsdnProtocolActions is UsdnProtocolStorage, IUsdnProtocolActio
     /// @inheritdoc IUsdnProtocolActions
     function initiateDeposit(
         uint128 amount,
-        uint128 amountMinOut,
+        uint256 amountMinOut,
         address to,
         address payable validator,
         Permit2TokenBitfield.Bitfield permit2TokenBitfield,
