+++ test/unit/UsdnProtocol/utils/Handler.sol
@@ -710,7 +710,7 @@ contract UsdnProtocolHandler is UsdnProtocolImpl, Test {
     function i_prepareInitiateDepositData(
         address validator,
         uint128 amount,
-        uint128 amountMinOut,
+        uint256 amountMinOut,
         bytes calldata currentPriceData
     ) public returns (ActionsVault.InitiateDepositData memory data_) {
         return ActionsVault._prepareInitiateDepositData(s, validator, amount, amountMinOut, currentPriceData);
